introduce FrontendFeature
