+++ .changeset/big-ads-travel.md
@@ -1,0 +1,5 @@
+---
+'@backstage/frontend-plugin-api': patch
+---
+
+Add the `FrontendFeature` type, which is the union of `BackstagePlugin` and `ExtensionOverrides`
