+++ .changeset/brown-emus-explode.md
@@ -1,0 +1,6 @@
+---
+'@backstage/frontend-app-api': patch
+'@backstage/core-compat-api': patch
+---
+
+Leverage the new `FrontendFeature` type to simplify interfaces
