+++ packages/frontend-app-api/src/wiring/createApp.tsx
@@ -19,7 +19,6 @@ import { ConfigReader, Config } from '@backstage/config';
 import {
   AppTree,
   appTreeApiRef,
-  BackstagePlugin,
   ComponentRef,
   componentsApiRef,
   coreExtensionData,
@@ -29,7 +28,7 @@ import {
   createThemeExtension,
   createTranslationExtension,
   ExtensionDataRef,
-  ExtensionOverrides,
+  FrontendFeature,
   RouteRef,
   useRouteRef,
 } from '@backstage/frontend-plugin-api';
@@ -217,8 +216,8 @@ export function createExtensionTree(options: {
 }
 
 function deduplicateFeatures(
-  allFeatures: (BackstagePlugin | ExtensionOverrides)[],
-): (BackstagePlugin | ExtensionOverrides)[] {
+  allFeatures: FrontendFeature[],
+): FrontendFeature[] {
   // Start by removing duplicates by reference
   const features = Array.from(new Set(allFeatures));
 
@@ -241,12 +240,10 @@ function deduplicateFeatures(
 
 /** @public */
 export function createApp(options?: {
-  features?: (BackstagePlugin | ExtensionOverrides)[];
+  features?: FrontendFeature[];
   configLoader?: () => Promise<{ config: ConfigApi }>;
   bindRoutes?(context: { bind: CreateAppRouteBinder }): void;
-  featureLoader?: (ctx: {
-    config: ConfigApi;
-  }) => Promise<(BackstagePlugin | ExtensionOverrides)[]>;
+  featureLoader?: (ctx: { config: ConfigApi }) => Promise<FrontendFeature[]>;
 }): {
   createRoot(): JSX.Element;
 } {
@@ -291,7 +288,7 @@ export function createApp(options?: {
  * @public
  */
 export function createSpecializedApp(options?: {
-  features?: (BackstagePlugin | ExtensionOverrides)[];
+  features?: FrontendFeature[];
   config?: ConfigApi;
   bindRoutes?(context: { bind: CreateAppRouteBinder }): void;
 }): { createRoot(): JSX.Element } {
