+++ packages/frontend-plugin-api/src/wiring/createExtensionOverrides.ts
@@ -19,17 +19,12 @@ import {
   Extension,
   resolveExtensionDefinition,
 } from './resolveExtensionDefinition';
-import { FeatureFlagConfig } from './types';
+import { ExtensionOverrides, FeatureFlagConfig } from './types';
 
 /** @public */
 export interface ExtensionOverridesOptions {
   extensions: ExtensionDefinition<unknown>[];
   featureFlags?: FeatureFlagConfig[];
-}
-
-/** @public */
-export interface ExtensionOverrides {
-  readonly $$type: '@backstage/ExtensionOverrides';
 }
 
 /** @internal */
