+++ packages/frontend-plugin-api/src/wiring/createPlugin.ts
@@ -15,18 +15,16 @@
  */
 
 import { ExtensionDefinition } from './createExtension';
-import { ExternalRouteRef, RouteRef } from '../routing';
-import { FeatureFlagConfig } from './types';
 import {
   Extension,
   resolveExtensionDefinition,
 } from './resolveExtensionDefinition';
-
-/** @public */
-export type AnyRoutes = { [name in string]: RouteRef };
-
-/** @public */
-export type AnyExternalRoutes = { [name in string]: ExternalRouteRef };
+import {
+  AnyExternalRoutes,
+  AnyRoutes,
+  BackstagePlugin,
+  FeatureFlagConfig,
+} from './types';
 
 /** @public */
 export interface PluginOptions<
@@ -38,17 +36,6 @@ export interface PluginOptions<
   externalRoutes?: ExternalRoutes;
   extensions?: ExtensionDefinition<unknown>[];
   featureFlags?: FeatureFlagConfig[];
-}
-
-/** @public */
-export interface BackstagePlugin<
-  Routes extends AnyRoutes = AnyRoutes,
-  ExternalRoutes extends AnyExternalRoutes = AnyExternalRoutes,
-> {
-  readonly $$type: '@backstage/BackstagePlugin';
-  readonly id: string;
-  readonly routes: Routes;
-  readonly externalRoutes: ExternalRoutes;
 }
 
 /** @public */
