+++ packages/frontend-plugin-api/src/wiring/index.ts
@@ -34,17 +34,17 @@ export {
   type ExtensionDataRef,
   type ConfigurableExtensionDataRef,
 } from './createExtensionDataRef';
-export {
-  createPlugin,
-  type BackstagePlugin,
-  type PluginOptions,
-  type AnyRoutes,
-  type AnyExternalRoutes,
-} from './createPlugin';
+export { createPlugin, type PluginOptions } from './createPlugin';
 export {
   createExtensionOverrides,
-  type ExtensionOverrides,
   type ExtensionOverridesOptions,
 } from './createExtensionOverrides';
 export { type Extension } from './resolveExtensionDefinition';
-export type { FeatureFlagConfig } from './types';
+export {
+  type AnyRoutes,
+  type AnyExternalRoutes,
+  type BackstagePlugin,
+  type ExtensionOverrides,
+  type FeatureFlagConfig,
+  type FrontendFeature,
+} from './types';
